+++ src/models/Demos.ts
@@ -19,8 +19,11 @@ export interface Demo {
 export interface Step {
   action: Action;
   path: string;
+
   content?: string;
   contentPath?: string;
   position?: string | number;
   timeout?: number;
+  command?: string;
+  args?: any;
 }